requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning & Regression
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
statsmodels>=0.14.0

# Hyperparameter Optimization
optuna>=3.3.0
hyperopt>=0.2.7
scikit-optimize>=0.9.0

# Model Interpretation & Explainability
shap>=0.42.0
lime>=0.2.0.1
eli5>=0.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Statistical Analysis
scipy>=1.11.0

# Progress Bars & Utilities
tqdm>=4.66.0

# Data Validation
ydata-profiling>=4.0.0

# Time Series & Forecasting
prophet>=1.1.0
pmdarima>=2.0.0

# Marketing Mix Modeling & Bayesian Analysis
pymc>=5.0.0
arviz>=0.16.0

# API Connectors for Digital Media Platforms
google-ads>=22.0.0
facebook-business>=18.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Data Connectors & Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Excel & Reporting
xlsxwriter>=3.1.0
python-pptx>=0.6.21

# Feature Engineering & Preprocessing
category-encoders>=2.6.0
imbalanced-learn>=0.11.0

# Adstock & Transformation Libraries
numba>=0.58.0

# Jupyter Support
jupyter>=1.0.0
ipywidgets>=8.1.0

# Model Persistence & Tracking
joblib>=1.3.0
mlflow>=2.8.0

# Cross-validation & Model Selection
mlxtend>=0.23.0